feat: change model for posting comment function
- move Comment table in place/models.py to social/models.py
- makemigrations -> migrate
