+++ social/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 
-from social.models import Relationship
+from social.models import Relationship, Comment
 
 admin.site.register(Relationship)
+admin.site.register(Comment)
